+++ src/components/layout/SectionBody.tsx
@@ -1,6 +1,8 @@
-import { SectionPaddingSize } from "components/layout/Section"
+import {
+  SectionPaddingSize,
+  sectionPartPadding,
+} from "components/layout/Section"
 import styled, { css } from "styled-components"
-import { SPACINGS } from "styles/spacings"
 import { CSSProp } from "styles/types"
 
 export type SectionBodyProps = {
@@ -18,19 +20,7 @@ export const SectionBody = styled.div<SectionBodyProps>`
   width: 100%;
   flex: 1;
   overflow-y: ${({ scrollable }) => (scrollable ? "scroll" : "unset")};
-  ${({ paddingSize, theme }) => {
-    if (paddingSize === "small")
-      return css`
-        padding: ${SPACINGS.SM};
-      `
-    if (paddingSize === "medium")
-      return css`
-        padding: ${SPACINGS.MD};
-      `
-    return css`
-      padding: ${theme.styles.section.body.padding};
-    `
-  }}
+  ${({ paddingSize }) => sectionPartPadding({ paddingSize, part: "body" })}
   ${({ noHorizontalPadding }) =>
     noHorizontalPadding ? removeHorizontalPadding : ""};
   ${({ _css }: CSSProp) => _css}